fix: font display is irregular for WeaselDeployer.
The current font is set to “MS Shell Dlg”, which is mapped to “SimSun” on Chinese operating systems instead of the system default font, which throws to very terrible display results.

I just changed it to "Segoe UI" and the problem was solved.

Q: Why not set it to "Microsoft YaHei" or "Microsoft JhengHei" immediately?

A: Since many Chinese users have both fonts on their computers, doing so will most likely result in them being forced to be mapped to the wrong font.

Q: Why Segoe UI?

A: Because it's Windows's default font for the English version, than can be a fallback to correct fonts for other languages referenced by system fontlink.